feat: category 변경에 따른 투두 변경
